// src/restrorealm/src/app/shared/enum/order-status.enum.ts
export enum OrderStatus {
    PENDING = 'PENDING',
    IN_PROGRESS = 'IN_PROGRESS',
    CONFIRMED = 'CONFIRMED',
    PREPARING = 'PREPARING',
    READY_FOR_PICKUP = 'READY_FOR_PICKUP',
    COMPLETED = 'COMPLETED',
    CANCELLED = 'CANCELLED',
    REJECTED = 'REJECTED',
    PAYMENT_PENDING = 'PAYMENT_PENDING',
    PAYMENT_SUCCESSFUL = 'PAYMENT_SUCCESSFUL',
    PAYMENT_FAILED = 'PAYMENT_FAILED',
    PAYMENT_REFUNDED = 'PAYMENT_REFUNDED',
    CASHIER_PENDING = 'CASHIER_PENDING'
}